Expense

+++ backend/controllers/inventory.controller.js
@@ -58,6 +58,8 @@ export const editInventory = async (req, res) => {
 export const deleteInventory = async (req, res) => {
     try {
         const { id } = req.params;
+        const find = await InventoryModal.findById(id);
+        if (!find) return res.status(404).json({ message: "Inventory not found" });
         await InventoryModal.findByIdAndDelete(id);
         res.status(200)
             .json({
